Make legend more d3 component-like
2 types of legend - colour (categories) and symbol (shape and colour)
Moved filter functions to their own module
