+++ packages/perspective-viewer-d3fc/src/js/series/pointSeries.js
@@ -9,7 +9,8 @@
 import * as d3 from "d3";
 import * as fc from "d3fc";
 import {tooltip} from "../tooltip/tooltip";
-import {groupFromKey} from "../series/seriesKey";
+import {groupFromKey} from "./seriesKey";
+import {withoutOpacity} from "./seriesColours";
 
 const symbols = [d3.symbolCircle, d3.symbolCross, d3.symbolDiamond, d3.symbolSquare, d3.symbolStar, d3.symbolTriangle, d3.symbolWye];
 
@@ -29,16 +30,11 @@ export function pointSeries(settings, seriesKey, size, colour, symbols) {
     series.decorate(selection => {
         tooltip()(selection, settings);
         if (colour) {
-            selection.style("stroke", () => withOutOpacity(colour(seriesKey))).style("fill", () => colour(seriesKey));
+            selection.style("stroke", () => withoutOpacity(colour(seriesKey))).style("fill", () => colour(seriesKey));
         }
     });
 
     return series;
-}
-
-function withOutOpacity(colour) {
-    const lastComma = colour.lastIndexOf(",");
-    return lastComma !== -1 ? `${colour.substring(0, lastComma)})` : colour;
 }
 
 export function symbolTypeFromGroups(settings) {
